+++ app/avo/cards/google_calendar.rb
@@ -1,0 +1,8 @@
+class Avo::Cards::GoogleCalendar < Avo::Cards::PartialCard
+  self.id = "google_calendar"
+  self.label = "Google calendar"
+  self.partial = "avo/cards/google_calendar"
+  # self.display_header = true
+  # self.cols = 1
+  # self.rows = 1
+end