+++ app/avo/cards/recent_programs.rb
@@ -1,0 +1,8 @@
+class Avo::Cards::RecentPrograms < Avo::Cards::PartialCard
+  self.id = "recent_programs"
+  self.label = "Recent programs"
+  self.partial = "avo/cards/recent_programs"
+  # self.display_header = true
+  # self.cols = 1
+  # self.rows = 1
+end