+++ app/models/daily_metric.rb
@@ -1,9 +1,15 @@
 class DailyMetric < ApplicationRecord
   belongs_to :user
+  belongs_to :user_dietary_plan, optional: true
 
+  before_validation :assign_to_active_plan, on: :create
   before_save :parse_content_with_ai, if: :should_parse_ai?
 
   private
+
+  def assign_to_active_plan
+    self.user_dietary_plan ||= user.user_dietary_plans.active.last
+  end
 
   def should_parse_ai?
     raw_message_content.present? && (raw_message_content_changed? || ai_parsed_json.blank?)